backend/requirements.txt
chromadb
sentence-transformers
google-generativeai
python-dotenv
urllib3<2
fastapi
uvicorn
python-multipart
